fix(runtime): IsaCache 槽位哈希在 u64 上混合，修 wasm32 常量越界
CI package-wasm：`0x9E37_79B9_7F4A_7C15` 对 32 位 usize 越界。改为 u64 混合后截断到
usize；`cargo check --target wasm32-unknown-unknown` 通过，isa_cache 单测通过，xtask test 全绿。

+++ src/runtime/src/vm_context/isa_cache.rs
@@ -59,9 +59,10 @@ impl IsaCache {
     #[inline]
     pub(crate) fn index(td: usize, target: usize) -> usize {
         // Both keys are pointers (16-byte aligned heap allocations → low bits are zero);
-        // fold the informative bits of each before masking.
-        let h = (td >> 4).wrapping_mul(0x9E37_79B9_7F4A_7C15) ^ (target >> 3);
-        (h ^ (h >> 17)) & (SLOTS - 1)
+        // fold the informative bits of each before masking. Mix in u64 so the constant is
+        // valid on 32-bit targets too (wasm32: `usize` is 32 bits).
+        let h = ((td as u64) >> 4).wrapping_mul(0x9E37_79B9_7F4A_7C15) ^ ((target as u64) >> 3);
+        ((h ^ (h >> 17)) as usize) & (SLOTS - 1)
     }
 
     /// Cached verdict for (`td`, `target`), if this exact pair was installed.
